extract HtmlView into components/
